BOM upload autocreate: Add Team to ACL
